+++ modules/circe/src/test/scala/io/bartholomews/fsclient/samples/v1/TokenSignatureExample.scala
@@ -1,12 +1,11 @@
 package io.bartholomews.fsclient.samples.v1
 
 object TokenSignatureExample extends App {
-  import io.bartholomews.fsclient.client.ClientData.sampleRedirectUri
   import io.bartholomews.fsclient.core.config.UserAgent
   import io.bartholomews.fsclient.core.oauth.v1.OAuthV1.{Consumer, SignatureMethod}
   import io.bartholomews.fsclient.core.oauth.v1.TemporaryCredentials
-  import io.bartholomews.fsclient.core.oauth.v2.OAuthV2.RedirectUri
   import io.bartholomews.fsclient.core.oauth.{
+    RedirectUri,
     RequestTokenCredentials,
     ResourceOwnerAuthorizationUri,
     TemporaryCredentialsRequest
@@ -61,7 +60,7 @@ object TokenSignatureExample extends App {
 
   // a successful `resourceOwnerAuthorizationUriResponse` will have the token in the query parameters:
   val resourceOwnerAuthorizationUriResponse =
-    sampleRedirectUri.value.withParams(("oauth_token", "AAA"), ("oauth_verifier", "ZZZ"))
+    myRedirectUri.value.withParams(("oauth_token", "AAA"), ("oauth_verifier", "ZZZ"))
 
   // 3. Extract the Token Credentials
   val maybeRequestTokenCredentials: Either[ResponseException[String, Exception], RequestTokenCredentials] =