+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `AccessTokenSigner` using `currentTimeMillis` instead of `nanoTime`
 - `baseRequest` taking an `userAgent` instead of `client`
 - `signer` needs to be passed explicitly
+- `RedirectUri` moved up from `oauth.v2` to `oauth` package
 ### Fixed
 
 ## [0.1.1] - 2021-01-30